requirements.txt
opencv-python>=4.10.0
numpy>=1.23.0,<2.0.0
scipy>=1.11.0
torch>=2.1.0
torchvision>=0.16.0
open-rppg>=0.1.0
hsemotion>=0.3.0
timm>=0.9.0,<1.0.0
matplotlib>=3.7.0
